refactor code && change critic network structure
