+++ ddpg/Noise.py
@@ -5,10 +5,16 @@
 # ===========================
 import numpy as np
 
-class Noise(object):
+
+class OrnsteinUhlenbeckNoise(object):
+
 
     def __init__(self, delta=0.5, sigma=0.5, ou_a=3.0, ou_mu=0.0):
         # Noise parameters
+        # DELTA = 0.1 # The rate of change (time)
+        # SIGMA = 5 # Volatility of the stochastic processes
+        # OU_A = 3. # The rate of mean reversion
+        # OU_MU = 0. # The long run average interest rate
         self.delta = delta
         self.sigma = sigma
         self.ou_a = ou_a
